timestamp: add support for sub-second input
sub-second values are stripped off and ignored, but they are now
processed correctly.  For example, the input "1499358871000" is now
correctly processed as a date in the year 2017, not 49482.

This does have the side effect of not being able to handle unix
timestamps after 9999999999 (2286-11-20).  For how I use this package,
I'm not too worried about that, but this could be made a little smarter
if needed.